Use parseFloat instead of parseInt

+++ src/Containers/AutomationCalculator/AutomationCalculator.js
@@ -442,7 +442,7 @@ const AutomationCalculator = () => {
                                           step="any"
                                           min="0"
                                           aria-label="manual-cost"
-                                          value={ parseInt(costManual) }
+                                          value={ parseFloat(costManual) }
                                           onChange={ e => setCostManual(e) }
                                       />
                                       <InputGroupText>/hr</InputGroupText>
@@ -460,7 +460,7 @@ const AutomationCalculator = () => {
                                           step="any"
                                           min="0"
                                           aria-label="automation-cost"
-                                          value={ parseInt(costAutomation) }
+                                          value={ parseFloat(costAutomation) }
                                           onChange={ e => setCostAutomation(e) }
                                       />
                                       <InputGroupText>/hr</InputGroupText>
